pickup place, basic food
important: interaction change.
reference reservoir and counter class to understand how it functions.
if an object is interactable, it needs to use the Interactable namespace.
Then use a OnInteract function to handle input from the player.

Other classes:
food "reservoir" is infinite avaliable food to pickup/ place back into correct reservoir. Food class is a ScriptableObject which needs more work done on it